Update superstructuremotion
Planner should basically do this now:

```
Figure out if above crossbar and am trying to hit said crossbar (i.e. proximal beyond -90), if so move elevator if necessary to avoid crossbar and then put it in stowed

figure out the lowest point possible attainable in the set of numbers in proximal and wrist angles (so like pointing as far down as possible and slaming the elevator down, then trying to move the intake, or trying to move the intake without moving the elevator or anything. This is ANDed with a is within frame perimeter so that we are allowed to retrieve cargo. If so, move the elevator up just enough to make it safe

move the intake to it's target position, followed by the elevator
```